+++ git_key.pub
@@ -1,0 +1,1 @@
+ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQCtob6/A6io9LdEwI7gAgc8TMyoEHx7GXSg9yNu9btQT3Mz0Jy26hNiZVBTiD0jOsxKAIV2T7Uq1pcNMioqZclZfhLpqMtpE0jatHCvFyiVoUc/tFnl3VD6aM6QdHe2Fo/6qINykj9z97tf5obXiVJ0KCbWyaCsvvTm5DhXGtbfbc9PqOLUnELVg2v0xzBuwh8Wl7snImVHzsbbWlp2zST5fCxA3BDR48omH4U0YCFfwyu+rpqzMFxBSH8SrOSEI5qouAzaeLGs0GcoQSBjyrAdbe/Hq+1Tb8pnKgR5eKwycz1rCVlZ9wPLOclElo5yy0RTjZdl5Os4BcCHWF3yGc46cNg4E0OJ20thjJ4mB2CZjr68LGdRI8JwCrd+sIzjLn+yZMbRd8cOSo0PM+n6EtOwEPWp4nfbj8sph4aZzyR3VJSIenxgzo30LBKhrs+V0ppNU608h19qcEl/QsL12Y89LykQoaN3rJxWb+QbOQFUQ1l6o1b+vRwDg1F5GyX/XJk= hp@DESKTOP-HBDLLA9
